device_farm_v5/requirements.txt
# Device Farm v5 - Production Requirements
# Core Dependencies
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.13.0
sqlite3

# Device Control and Automation
pure-python-adb>=0.3.0.dev0
Appium-Python-Client>=3.1.0
selenium>=4.15.0
uiautomator2>=2.16.25

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-socketio>=5.9.0

# Async and Concurrency  
asyncio
aiohttp>=3.9.0
aiofiles>=23.2.0

# Logging and Monitoring
loguru>=0.7.2
structlog>=23.2.0

# Configuration and Environment
python-decouple>=3.8
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Task Queue and Job Management
celery[redis]>=5.3.4
redis>=5.0.1
kombu>=5.3.4

# Security and Authentication
cryptography>=41.0.7
jwt>=1.3.1
bcrypt>=4.1.2

# Testing and Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
httpx>=0.25.2

# API Client Helpers
httpx>=0.25.2
tenacity>=8.2.3
backoff>=2.2.1

# Performance and Monitoring
psutil>=5.9.6
py-cpuinfo>=9.0.0
GPUtil>=1.4.0

# Data Processing
pandas>=2.1.3
numpy>=1.25.2

# Network and Proxy
requests[socks]>=2.31.0
PySocks>=1.7.1

# Development Tools
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
flake8>=6.1.0